Alpha Release.  Requires updates to bf-py before fully functional

+++ bf_algo_sar.py
@@ -11,20 +11,20 @@ try:
 except:
     from osgeo import gdal, osr
 
-def main(img_path, out_path, kernelSize=10, scaleFactor=0.1):
+def main(img_path, out_path, kernelSize=10, scaleFactor=0.1, cleanup=1):
     #run bf-Sar2Shoreline --> returns binary mask
     name = bf_Sar2Shoreline.xO_names(out_path)
-    binary_mask ='%s/%s_binary.%s' % (name['directory'],
-                                  name['basename'],
-                                  name['extension'])
+    binary_mask ='%s/%s_binary.tif' % (name['directory'],
+                                  name['basename'])
     print binary_mask
     status = bf_Sar2Shoreline.WaterExtractionSAR(img_path,
                                                  out_path=binary_mask,
                                                  kernelSize=kernelSize,
-                                                 scaleFactor=sf)
+                                                 scaleFactor=scaleFactor)
     #run vectorizing code from bf-py --> returns vector
     coastline = trace_it(binary_mask)
-    os.remove(binary_mask)
+    if cleanup == 1:
+        os.remove(binary_mask)
     # write to file
     with open(out_path, 'w') as out_path:
         json.dump(coastline, out_path)
@@ -45,6 +45,7 @@ if __name__ == '__main__':
     out_path = None
     kernelSize = 10
     sf = 0.1
+    cleanup = 0
 
     for i in range(len(sys.argv)-1):
         arg = sys.argv[i]
@@ -56,11 +57,13 @@ if __name__ == '__main__':
             kernelSize = int(sys.argv[i+1])
         elif arg == '-s':
             sf = float(sys.argv[i+1])
+        elif arg == '-c':
+            cleanup == int(sys.argv[i+1])
 
     if img_path is None:
         usage()
     if out_path is None:
         usage()
 
-    main(img_path, out_path, kernelSize=kernelSize, scaleFactor=sf)
+    main(img_path, out_path, kernelSize=kernelSize, scaleFactor=sf, cleanup=cleanup)
     sys.exit(0)
